refactor: reorganize drizzle parser for multi-database support
Restructure drizzle parser directory to support multiple databases:
- Move PostgreSQL-specific files to postgres/ subdirectory
- Organize tests in __tests__/ subdirectories
- Update import paths throughout the codebase
- Prepare structure for future MySQL/SQLite support
