+++ frontend/packages/db-structure/src/parser/drizzle/index.ts
@@ -1,1 +1,3 @@
-export { processor } from './mainParser.js'
+// Currently only PostgreSQL is supported
+// In the future, we can add support for MySQL, SQLite, etc.
+export { processor } from './postgres/index.js'
