+++ frontend/packages/db-structure/src/parser/drizzle/postgres/index.ts
@@ -1,0 +1,1 @@
+export { processor } from './mainParser.js'
