+++ frontend/packages/db-structure/src/parser/drizzle/postgres/mainParser.ts
@@ -4,7 +4,7 @@
 
 import type { Module, VariableDeclarator } from '@swc/core'
 import { parseSync } from '@swc/core'
-import type { Processor, ProcessResult } from '../types.js'
+import type { Processor, ProcessResult } from '../../types.js'
 import { isPgTableCall, isSchemaTableCall } from './astUtils.js'
 import { convertDrizzleTablesToInternal } from './converter.js'
 import { parsePgEnumCall } from './enumParser.js'
@@ -161,4 +161,5 @@ const parseDrizzleSchemaString = (
   }
 }
 
-export const processor: Processor = (str) => parseDrizzleSchemaString(str)
+export const processor: Processor = (str: string) =>
+  parseDrizzleSchemaString(str)
